+++ status.php
@@ -1,5 +1,6 @@
 <?php
 include "./config.php";
+include "./utils.php";
 
 $force_login_redirect = true;
 include "./authentication.php";
@@ -19,9 +20,10 @@ include "./authentication.php";
     </head>
     <body class="inlinebody">
         <?php
+        $instance_config = load_instance_config($config);
 
-        $heartbeat_file_path = $config["interface_directory"] . "/heartbeat.json";
-        if (is_dir($config["interface_directory"]) == true) { // Check to make sure the specified interface directory exists.
+        $heartbeat_file_path = $instance_config["general"]["interface_directory"] . "/heartbeat.json";
+        if (is_dir($instance_config["general"]["interface_directory"]) == true) { // Check to make sure the specified interface directory exists.
             if (file_exists($heartbeat_file_path)) { // Check to see if the heartbeat file exists.
                 $heartbeat_log = json_decode(file_get_contents($heartbeat_file_path), true); // Load the heartbeat file from JSON data.
             } else { // If the heartbeat file doesn't exist, then load a blank placeholder instead.